Update logic for fetching and testing url
The logic to retrieve the url had to be altered as Express treats a /
as a delimeter for a new parameter. A Regex was written that fetches
everything after /new/. Additionally the logic for testing if the url
is valid was adjusted to use the valid-url module.

Some precursor logic was added to test if the url starts with www. In
this case the server prepends http:// to the url before passing it to
the test function. If the url is valid, it currently responds that it
is a valid url. Functionality for MongoDB still needs to be added. If
it is not valid an object is included in the response with the original
url and an error message.